requirements.txt
cycler==0.10.0
numpy==1.18.1
joblib==0.16.0
kiwisolver==1.2.0
pydub==0.23.1
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.21.3
scipy==1.4.1
six==1.15.0
tqdm==4.44.1
telegram
python-telegram-bot==11.1.0
SpeechRecognition==3.8.1
Pillow
